+++ CHANGES.rst
@@ -42,6 +42,7 @@ Release History
 - Fixed an error that was thrown when calling ``get_tensor`` on a ``Signal``
   that was first initialized inside the Simulation while loop
   (`#56 <https://github.com/nengo/nengo-dl/issues/56>`_)
+- Allow TensorNodes to run in Nengo GUI.
 
 1.2.0 (September 5, 2018)
 -------------------------